add function ft_bzero and main ft_bzero and modif ft_memset

+++ ft_memset.c
@@ -1,6 +1,8 @@
-void	*ft_memset(void *s, int c, unsigned int n)
+#include <stddef.h>
+
+void	*ft_memset(void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	char			*tab;
 
 	tab = s;
